+++ src/main/java/com/umc/banddy/domain/chat/web/controller/ChatContoller.java
@@ -4,9 +4,11 @@ import com.umc.banddy.domain.chat.domain.ChatRoom;
 import com.umc.banddy.domain.chat.service.ChatMessageService;
 import com.umc.banddy.domain.chat.service.ChatRoomService;
 import com.umc.banddy.domain.chat.service.ChatService;
-import com.umc.banddy.domain.chat.web.dto.ChatRoom.*;
-import com.umc.banddy.domain.chat.web.dto.Message.ChatSystemResponse;
-import com.umc.banddy.domain.chat.web.dto.Message.CursorChatMessageResponse;
+import com.umc.banddy.domain.chat.web.dto.chatRoom.*;
+import com.umc.banddy.domain.chat.web.dto.chatRoom.creation.*;
+import com.umc.banddy.domain.chat.web.dto.chatRoom.roomList.ChatRoomListResponse;
+import com.umc.banddy.domain.chat.web.dto.message.ChatSystemResponse;
+import com.umc.banddy.domain.chat.web.dto.message.CursorChatMessageResponse;
 import com.umc.banddy.domain.member.domain.Member;
 import com.umc.banddy.global.security.jwt.JwtTokenUtil;
 import io.swagger.v3.oas.annotations.Operation;
