+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/creation/BandJoinRequest.java
@@ -1,4 +1,4 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.creation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @Builder
-public class PrivateChatRoomRequest {
+public class BandJoinRequest {
 
-    private Long memberId;
+    private String session;
 }
