+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/creation/GroupChatRoomResponse.java
@@ -1,11 +1,10 @@
- package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+ package com.umc.banddy.domain.chat.web.dto.chatRoom.creation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.springframework.web.multipart.MultipartFile;
 
-@Getter
+ @Getter
 @AllArgsConstructor
 @Builder
 public class GroupChatRoomResponse {
