+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/creation/PrivateChatRoomRequest.java
@@ -1,4 +1,4 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.creation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @Builder
-public class PrivateChatRoomResponse {
+public class PrivateChatRoomRequest {
 
-    private Long roomId;
+    private Long memberId;
 }
