+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/roomList/AppliedRoomInfo.java
@@ -1,4 +1,4 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.roomList;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
