+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/roomList/ChatRoomInfo.java
@@ -1,12 +1,9 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.roomList;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @SuperBuilder
