+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/roomList/ChatRoomInfoDto.java
@@ -1,7 +1,6 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.roomList;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
+import com.umc.banddy.domain.chat.web.dto.chatRoom.MemberInfo;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 
@@ -9,7 +8,7 @@ import java.util.List;
 
 @Getter
 @SuperBuilder
-public class ChatRoomInfoDto extends ChatRoomInfo{
+public class ChatRoomInfoDto extends ChatRoomInfo {
 
     private Long roomId;
     private List<MemberInfo> memberInfos;
