+++ src/main/java/com/umc/banddy/domain/chat/web/dto/chatRoom/roomList/ChatRoomListResponse.java
@@ -1,4 +1,4 @@
-package com.umc.banddy.domain.chat.web.dto.ChatRoom;
+package com.umc.banddy.domain.chat.web.dto.chatRoom.roomList;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,8 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 @Builder
-public class FriendsChatRoomResponse {
-    private List<FriendChatRoom> rooms;
+public class ChatRoomListResponse {
+
+    private List<ChatRoomInfo> chatRoomInfos;
+
 }
